refactor(studio): split viewer shaders.rs into per-pass submodules
Break the 1150-line shaders.rs into a shaders/ module (mesh, grid,
shadow, imposter, gizmo). The root re-exports flat so call sites keep
using shaders::VS_SRC etc. unchanged.
